fix: leaving off with internal error inside internal error
Something isn't working right with the way requests are routed and suspect the mix of cloudflare pages build dir and the worker functions. At this point might be easier to clone and setup functionality in an actual worker and serve the API from the root. Can provide generated docs similar to past projects too.